Сделал рефактор утилит
