+++ src/main.js
@@ -9,11 +9,8 @@ import ShowMoreButtonView from './view/films/show-more-button.js';
 import FilmListExtraView from './view/films/films-list-extra.js';
 import FilmModalView from './view/films/film-modal.js';
 import { getTestData } from './mock/films.js';
-import {
-  getRandomPositiveInteger,
-  InsertPosition,
-  renderElement
-} from './utils/utils.js';
+import { InsertPosition, renderElement } from './utils/dom';
+import { getRandomPositiveInteger } from './utils/test-data.js';
 
 const FILMS_PER_ROW = 5;
 
@@ -52,7 +49,7 @@ const openFilmModal = (filmData) => {
   }
 
   filmModal = new FilmModalView(filmData);
-  renderElement(document.body, filmModal.getElement(), InsertPosition.BEFORE_END);
+  renderElement(document.body, filmModal, InsertPosition.BEFORE_END);
 
   document.addEventListener('keydown', onEscKeyDown);
   filmModal.getElement().querySelector('.film-details__close-btn').addEventListener('click', () => onCloseClick());
@@ -66,10 +63,10 @@ const registerModalOpenListeners = (elements, filmData) => {
 
 
 const renderStatsPage = () => {
-  renderElement(headerElement, new UserProfileView(filterData.historyList.length).getElement(), InsertPosition.BEFORE_END);
-  renderElement(mainElement, new SiteMenuView(filterData).getElement(), InsertPosition.AFTER_BEGIN);
-  renderElement(mainElement, new StatisticView(films).getElement(), InsertPosition.BEFORE_END);
-  renderElement(footerElement, new FooterStatisticView(filmsCount).getElement(), InsertPosition.BEFORE_END);
+  renderElement(headerElement, new UserProfileView(filterData.historyList.length), InsertPosition.BEFORE_END);
+  renderElement(mainElement, new SiteMenuView(filterData), InsertPosition.AFTER_BEGIN);
+  renderElement(mainElement, new StatisticView(films), InsertPosition.BEFORE_END);
+  renderElement(footerElement, new FooterStatisticView(filmsCount), InsertPosition.BEFORE_END);
 };
 
 const renderMainPage = () => {
@@ -81,7 +78,7 @@ const renderMainPage = () => {
     return () => {
       for (lastIndex; lastIndex < limit; lastIndex++) {
         const filmCard = new FilmCardView(films[lastIndex]);
-        renderElement(containerElement, filmCard.getElement(), InsertPosition.BEFORE_END);
+        renderElement(containerElement, filmCard, InsertPosition.BEFORE_END);
 
         const modalTriggers = filmCard.getElement().querySelectorAll('.film-card__title, .film-card__poster, .film-card__comments');
 
@@ -94,17 +91,17 @@ const renderMainPage = () => {
   };
 
   //User profile
-  renderElement(headerElement, new UserProfileView(filterData.watchList.length).getElement(), InsertPosition.BEFORE_END);
+  renderElement(headerElement, new UserProfileView(filterData.watchList.length), InsertPosition.BEFORE_END);
 
   //Site nav
-  renderElement(mainElement, new SiteMenuView(filterData).getElement(), InsertPosition.AFTER_BEGIN);
+  renderElement(mainElement, new SiteMenuView(filterData), InsertPosition.AFTER_BEGIN);
 
   //Sort list
-  renderElement(mainElement, new SortListView().getElement(), InsertPosition.BEFORE_END);
+  renderElement(mainElement, new SortListView(), InsertPosition.BEFORE_END);
 
   //Films container
   const filmsList = new FilmListView();
-  renderElement(mainElement, filmsList.getElement(), InsertPosition.BEFORE_END);
+  renderElement(mainElement, filmsList, InsertPosition.BEFORE_END);
 
   if (films.length > 0) {
     const addFiveFilms = makeRenderFilmsFunction(filmsList.getFilmContainer());
@@ -130,13 +127,13 @@ const renderMainPage = () => {
 
 
   // Extra lists
-  renderElement(filmsList.getElement(), new FilmListExtraView().getElement(), InsertPosition.BEFORE_END);
-  renderElement(filmsList.getElement(), new FilmListExtraView().getElement(), InsertPosition.BEFORE_END);
+  renderElement(filmsList, new FilmListExtraView(), InsertPosition.BEFORE_END);
+  renderElement(filmsList, new FilmListExtraView(), InsertPosition.BEFORE_END);
 
   mainElement.appendChild(filmsList.getElement());
 
   //Footer statistics
-  renderElement(footerElement, new FooterStatisticView(filmsCount).getElement(), InsertPosition.BEFORE_END);
+  renderElement(footerElement, new FooterStatisticView(filmsCount), InsertPosition.BEFORE_END);
 };
 
 const setPage = (isStatPage = false) => isStatPage ? renderStatsPage() : renderMainPage();
